requirements/prod.txt
boto3>=1.3.0,<2
click>=7.0,<9.0
colorama>=0.2.5,<0.4.4
cfn-flip>=1.2.3,<2.0
deepdiff>=5.5.0,<6.0
Jinja2>=2.8,<3
jsonschema>=3.2,<3.3
networkx>=2.4,<2.6
packaging>=16.8,<17.0
PyYaml>=5.1,<6.0
sceptre-cmd-resolver>=1.1.3,<2
sceptre-file-resolver>=1.0.4,<2
six>=1.11.0,<2.0.0
tabulate>=0.8.9,<1.0.0
troposphere>=2.7,<3
